LayoutTests/imported/w3c/web-platform-tests/referrer-policy/gen/srcdoc-inherit.http-rp/origin/xhr.http-expected.txt

Blocked access to external URL http://www1.localhost:8800/common/security-features/subresource/xhr.py?redirection=keep-origin&action=purge&key=134849e3-f577-44ec-b1b6-4fe59a82f77d&path=%2Fmixed-content
CONSOLE MESSAGE: XMLHttpRequest cannot load http://www1.localhost:8800/common/security-features/subresource/xhr.py?redirection=keep-origin&action=purge&key=134849e3-f577-44ec-b1b6-4fe59a82f77d&path=%2Fmixed-content due to access control checks.
Blocked access to external URL http://www1.localhost:8800/common/security-features/subresource/xhr.py?redirection=no-redirect&action=purge&key=fc85eda2-cfd7-4e28-8e30-4b4bc1145daa&path=%2Fmixed-content
CONSOLE MESSAGE: XMLHttpRequest cannot load http://www1.localhost:8800/common/security-features/subresource/xhr.py?redirection=no-redirect&action=purge&key=fc85eda2-cfd7-4e28-8e30-4b4bc1145daa&path=%2Fmixed-content due to access control checks.
Blocked access to external URL http://www1.localhost:8800/common/security-features/subresource/xhr.py?redirection=swap-origin&action=purge&key=ddb6098e-a6d3-4fcb-a80b-f0b26e4f886b&path=%2Fmixed-content
CONSOLE MESSAGE: XMLHttpRequest cannot load http://www1.localhost:8800/common/security-features/subresource/xhr.py?redirection=swap-origin&action=purge&key=ddb6098e-a6d3-4fcb-a80b-f0b26e4f886b&path=%2Fmixed-content due to access control checks.
Blocked access to external URL https://www1.localhost:9443/common/security-features/subresource/xhr.py?redirection=keep-origin&action=purge&key=17f90991-485b-405e-9124-b2ac15112c7e&path=%2Fmixed-content
CONSOLE MESSAGE: XMLHttpRequest cannot load https://www1.localhost:9443/common/security-features/subresource/xhr.py?redirection=keep-origin&action=purge&key=17f90991-485b-405e-9124-b2ac15112c7e&path=%2Fmixed-content due to access control checks.
Blocked access to external URL https://www1.localhost:9443/common/security-features/subresource/xhr.py?redirection=no-redirect&action=purge&key=ee57c39f-f9b0-49f2-ac9e-51434e4b64e6&path=%2Fmixed-content
CONSOLE MESSAGE: XMLHttpRequest cannot load https://www1.localhost:9443/common/security-features/subresource/xhr.py?redirection=no-redirect&action=purge&key=ee57c39f-f9b0-49f2-ac9e-51434e4b64e6&path=%2Fmixed-content due to access control checks.
Blocked access to external URL https://www1.localhost:9443/common/security-features/subresource/xhr.py?redirection=swap-origin&action=purge&key=3845f5af-0db8-4152-a7dc-0eb35b7750a8&path=%2Fmixed-content
CONSOLE MESSAGE: XMLHttpRequest cannot load https://www1.localhost:9443/common/security-features/subresource/xhr.py?redirection=swap-origin&action=purge&key=3845f5af-0db8-4152-a7dc-0eb35b7750a8&path=%2Fmixed-content due to access control checks.
Blocked access to external URL http://www1.localhost:8800/common/security-features/subresource/xhr.py?action=purge&key=8453d8f2-76b0-4739-ba83-1ed9c5f2b2c8&path=%2Fmixed-content
Blocked access to external URL https://www1.localhost:9443/common/security-features/subresource/xhr.py?action=purge&key=16eff92a-69ed-47ba-8189-29c594fd9b8d&path=%2Fmixed-content

FAIL Referrer Policy: Expects origin for xhr to cross-http origin and keep-origin redirection from http context. promise_test: Unhandled rejection with value: "Network Error"
FAIL Referrer Policy: Expects origin for xhr to cross-http origin and no-redirect redirection from http context. promise_test: Unhandled rejection with value: "Network Error"
FAIL Referrer Policy: Expects origin for xhr to cross-http origin and swap-origin redirection from http context. promise_test: Unhandled rejection with value: "Network Error"
FAIL Referrer Policy: Expects origin for xhr to cross-https origin and keep-origin redirection from http context. promise_test: Unhandled rejection with value: "Network Error"
FAIL Referrer Policy: Expects origin for xhr to cross-https origin and no-redirect redirection from http context. promise_test: Unhandled rejection with value: "Network Error"
FAIL Referrer Policy: Expects origin for xhr to cross-https origin and swap-origin redirection from http context. promise_test: Unhandled rejection with value: "Network Error"
PASS Referrer Policy: Expects origin for xhr to same-http origin and keep-origin redirection from http context.
PASS Referrer Policy: Expects origin for xhr to same-http origin and no-redirect redirection from http context.
FAIL Referrer Policy: Expects origin for xhr to same-http origin and swap-origin redirection from http context. promise_test: Unhandled rejection with value: "Network Error"
PASS Referrer Policy: Expects origin for xhr to same-https origin and keep-origin redirection from http context.
PASS Referrer Policy: Expects origin for xhr to same-https origin and no-redirect redirection from http context.
FAIL Referrer Policy: Expects origin for xhr to same-https origin and swap-origin redirection from http context. promise_test: Unhandled rejection with value: "Network Error"
